Add a role to use vagrant insecure keys for root ssh
Vagrant creates a passwordless ssh key in
.vagrant/machines/<hostname>/virtualbox/private_key for each host.  By
default, public keys for these are set in vagrant user's authorized_keys
file.  This role also sets the public key in the
/root/.ssh/authorized_keys.  This is useful for using Higg's to generate
split installs, since you can point to the /vagrant/.vagrant/machines...
private_key's for the db and console hosts, and Higgs will be able to
log in as root and go about it's business.