(maint) Some example platform9 configuration
This is a sample platform9 configuration working for me locally.

The vms.yaml and openstack.yaml are checked in to record the
modifications I made to get vagrant to setup instances in platform9
using volume instances. The image can't be specified generally, but has
to be secified in a volume_boot hash, which will vary by 'size' assuming
that you change the volume size for different flavors.